+++ packages/player-react/src/index.ts
@@ -1,0 +1,6 @@
+export {
+  ensureQtiAssessmentItemPlayerDefined,
+  QtiAssessmentItemPlayer,
+  type QtiAssessmentItemPlayerHandle,
+  type QtiAssessmentItemPlayerProps,
+} from "./QtiAssessmentItemPlayer.js";